% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.entropy}
\alias{calculate.entropy}
\title{calculate.entropy}
\usage{
calculate.entropy(bam.file, out.folder = getwd(),
  out.name = "methclone")
}
\arguments{
\item{bam.file}{path to bam file containing the reads of the data set}

\item{out.folder}{folder to store the temporary file produced by methclone}

\item{out.name}{name of temporary file}
}
\value{
a data frame containing the position in the reference genome and the Entropy scores
}
\description{
This function calculates Methylation entropy by calling the methclone software to compute epiallele frequencies and
then uses the definition of entropy for the calculation.
}
