% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.epipolymorphism}
\alias{calculate.epipolymorphism}
\title{calculate.epipolymorphism}
\usage{
calculate.epipolymorphism(
  bam.file,
  out.folder = getwd(),
  out.name = "methclone"
)
}
\arguments{
\item{bam.file}{path to bam file containing the reads of the data set}

\item{out.folder}{folder to store the temporary file produced by methclone}

\item{out.name}{name of temporary file}
}
\value{
a data frame containing the position in the reference genome and the Epipolymorphism scores
}
\description{
This function calculates Epipolymorphism by calling the methclone software to compute epiallele frequencies and
then uses the definition of epipolymorphism for the calculation.
}
