% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.fdrp.by.chromosome}
\alias{calculate.fdrp.by.chromosome}
\title{calculate.fdrp.by.chromosome
This function calculates the FDRP scores for the reads given in the bam files
for the CpGs present in the annotation.}
\usage{
calculate.fdrp.by.chromosome(bam, anno, ignore.strand = TRUE)
}
\arguments{
\item{bam}{bam file with the reads from a bisulfite sequencing technique
already aligned to a reference genome}

\item{anno}{annotation as a GRanges object with the CpG sites to be analyzed}

\item{ignore.strand}{The \code{ignore.strand} parameter from the function \code{\link{findOverlaps}}}
}
\value{
vector of fdrp scores for the given CpG sites in anno
}
\description{
calculate.fdrp.by.chromosome
This function calculates the FDRP scores for the reads given in the bam files
for the CpGs present in the annotation.
}
\details{
This function is called by calculate.fdrps for each chromosome
			separately and calls toCpG as well as calculate.fdrp.site.
}
\author{
Michael Scherer
}
