% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.fdrp}
\alias{calculate.fdrp}
\title{calculate.fdrp}
\usage{
calculate.fdrp(bam_file, anno, path = getwd(), output_name, cores = 1,
  window.size = get.option("WINDOW.SIZE"))
}
\arguments{
\item{bam_file:}{bam-file of the sample to be analyzed}

\item{anno:}{CpG sites to be analyzed as a GRanges object}

\item{path:}{path to a folder where the FDRP CSV file should be written out}

\item{output_name:}{name of the output file}

\item{cores:}{number of cores available for the analysis}

\item{window.size:}{window size used to restrict the concordance/discordance classification of each read pair
DEFAULT: 50 as the maximum distance}
}
\value{
FDRP scores for the given annotation.
}
\description{
This function calculates the FDRPs for all CpG sites in
the annotation from the reads provided by the bam file.
}
\author{
Michael Scherer
}
