% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.mhl}
\alias{calculate.mhl}
\title{calculate.mhl}
\usage{
calculate.mhl(roi, bam.file, out.folder = getwd(),
  out.name = "mhl.txt", bam.type = "bismark")
}
\arguments{
\item{roi}{Region of interest for which haplotype information should be computed}

\item{bam.file}{path to bam file containing the reads}

\item{out.folder}{folder to write the output}

\item{out.name}{name of the output file}

\item{bam.type}{alignment tools used to create the bam file}
}
\value{
a data frame containing the position in the reference genome and the MHL scores
}
\description{
This function computed the Methylation Haplotype Load (MHL) for all sites in a given annotation by employing the scripts
downloaded from the paper's website
}
