% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.pdr.by.chromosome}
\alias{calculate.pdr.by.chromosome}
\title{calculate.pdr.by.chromosome}
\usage{
calculate.pdr.by.chromosome(bam, anno)
}
\arguments{
\item{bam}{bam file with the reads from a bisulfite sequencing technique
already aligned to a reference genome}

\item{anno}{annotation as a GRanges object with the CpG sites to be analyzed}
}
\value{
vector of pdr scores for the given CpG sites in anno
}
\description{
This function calculates the PDR scores for the reads given in the bam files
for the CpGs present in the annotation.
}
\author{
Michael Scherer
}
