% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.pdr}
\alias{calculate.pdr}
\title{calculate.pdr}
\usage{
calculate.pdr(
  bam.file,
  anno,
  log.path = getwd(),
  cores = 1,
  use.sex.chromosomes = FALSE,
  ignore.strand = TRUE
)
}
\arguments{
\item{bam.file}{bath to the bam file to be analyzed
already aligned to a reference genome}

\item{anno}{annotation as a GRanges object with the CpG sites to be analyzed}

\item{log.path}{location of the log file}

\item{cores}{number of cores available for the analysis}

\item{use.sex.chromosomes}{Flag indicating if scores are also to be computed for the sex chromosomes}

\item{ignore.strand}{The \code{ignore.strand} parameter from the function \code{\link{findOverlaps}}}
}
\value{
PDR scores for the given annotation.
}
\description{
This function calculates the PDRs for all analyzed CpG sites in
the bam file of the corresponding sample
}
\author{
Michael Scherer
}
