% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.pdr.site}
\alias{calculate.pdr.site}
\title{calculate.pdr.site
This function computes the PDR score for a given cpg site.}
\usage{
calculate.pdr.site(cpg, reads)
}
\arguments{
\item{pos:}{index of the cpg site}

\item{reads:}{GRanges object containing the reads needed for the calculation
of the PDR, already converted into the custom representation}
}
\value{
:			PDR score for the given CpG site
}
\description{
calculate.pdr.site
This function computes the PDR score for a given cpg site.
}
\author{
Michael Scherer
}
