% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.pdrs}
\alias{calculate.pdrs}
\title{calculate.pdr}
\usage{
calculate.pdrs(bam.file, anno, log.path = getwd(), cores = 1)
}
\arguments{
\item{bam.file:}{bath to the bam file to be analyzed
already aligned to a reference genome}

\item{anno:}{annotation as a GRanges object with the CpG sites to be analyzed}

\item{log.path:}{location of the log file}

\item{cores:}{number of cores available for the analysis}
}
\value{
PDR scores for the given annotation.
}
\description{
This function calculates the PDRs for all analyzed CpG sites in
the bam file of the corresponding sample
}
\author{
Michael Scherer
}
