% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.qfdrp}
\alias{calculate.qfdrp}
\title{calculate.qfdrp}
\usage{
calculate.qfdrp(bam_file, anno, path = getwd(), output_name, cores = 1,
  window.size = get.option("WINDOW.SIZE"))
}
\arguments{
\item{bam_file:}{bam file with the reads from a bisulfite sequencing technique
already aligned to a reference genome}

\item{anno:}{annotation as a GRanges object with the CpG sites to be analyzed}

\item{path:}{path to a folder where the FDRP CSV file should be written out}

\item{output_name:}{name of the output file}

\item{cores:}{number of cores available for the analysis}

\item{window.size:}{window size used to restrict the concordance/discordance classification of each read pair
DEFAULT: 50 as the maximum distance}
}
\value{
qFDRP scores for the given annotation.
}
\description{
This function calculates the qFDRP scores for the reads given in the bam files
for the CpGs present in the annotation.
}
\author{
Michael Scherer
}
