% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{compute.discordant}
\alias{compute.discordant}
\title{compute.discordant
This function decides for a given read pair if it is discordant or not.}
\usage{
compute.discordant(index, read1, read2, values, site)
}
\arguments{
\item{index}{index of the read pair}

\item{read1}{vector of indeces for the query reads for the read pairs}

\item{read2}{vector of indeces for the subject reads for the read pairs}

\item{values}{GRanges object containing the reads from which the read pairs
were calculated}

\item{site}{position of the CpG site of interest}
}
\value{
classification of the read pair as discordant or concordant
}
\description{
compute.discordant
This function decides for a given read pair if it is discordant or not.
}
\details{
The given read pair is only concordant, if all CpG sites that overlap between the
				reads reflect the same methylation status. Otherwise the read pair is classified
				as discordant.
}
\author{
Michael Scherer
}
