% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.score}
\alias{compute.score}
\title{compute.score}
\usage{
compute.score(bam.file, ..., score = "qfdrp")
}
\arguments{
\item{bam.file}{path to bam file containing the reads}

\item{...}{additional arugment. Either RnBSet, GRanges or empty (only for Epipolymorphism and Entropy)}

\item{score}{The ISH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}
}
\value{
data frame containing the annotation and the computed ISH scores
}
\description{
Generic function to call, passes its arugments either to \code{\link{compute.score.rnb}} or
\code{\link{compute.score.GRanges}}.
}
\examples{
\donttest{
load(system.file(file.path("extData","example_GRanges.RData"),package="ISH"))
example.rnb.set <- system.file(file.path("extData","small_rnbSet.zip"),package="ISH")
example.bam <- system.file(file.path("extData","small_example.bam"),package="ISH")
fdrp <- compute.score(bam.file=example.bam,example.GRanges,score="fdrp")
qfdrp <- compute.score(bam.file=example.bam,example.rnb.set,score="qfdrp")
}
}
\author{
Michael Scherer
}
