% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.score.rnb}
\alias{compute.score.rnb}
\title{compute.score.rnb}
\usage{
compute.score.rnb(bam.file, rnb.set, score)
}
\arguments{
\item{bam.file}{path to bam file containing the reads}

\item{rnb.set}{path to RnBSet contaning methylation, coverage and sample meta information}

\item{score}{The ISH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}
}
\value{
data frame containing the annotation and the computed ISH scores
}
\description{
Main function to compute ISH scores based on a input RnBSet and a bam file containing the reads.
}
\author{
Michael Scherer
}
