% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create.genomebrowser.track}
\alias{create.genomebrowser.track}
\title{create.genomebrowser.track}
\usage{
create.genomebrowser.track(score.output, sample.name = "Sample",
  store.path = getwd(), bin.width = 5000)
}
\arguments{
\item{score.output}{The output as generated through compute.score or compute.score.GRanges: a
data.frame with chromosome positions and the score values for those positions.}

\item{sample.name}{The name of the sample to be added to the file.}

\item{store.path}{The path to which the file is to be stored.}

\item{bin.width}{Width of bins in bp to aggregate the scores. If NULL, the scores are stored
as single CpGs.}
}
\description{
Converts the output of compute.score und compute.score.GRanges into genome browser tracks, that
can be directly imported into e.g. UCSC Genome Browser.
}
\author{
Michael Scherer
}
