% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ish.run.example}
\alias{ish.run.example}
\title{ish.run.example}
\usage{
ish.run.example(score = "qfdrp")
}
\arguments{
\item{score}{The ISH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}
}
\value{
A data frame containing the annotation and the corresponding score
}
\description{
Returns the scores for a small example data set of size 50kb
}
\examples{
\donttest{
ish.run.example("pdr")
}
}
\author{
Michael Scherer
}
