% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{restrict}
\alias{restrict}
\title{compute.discordant
This function restricts a read to those CpG sites that are at most WINDOW.SIZE away from one
another.}
\usage{
restrict(positions, cpg)
}
\arguments{
\item{positions}{positions of the CpG sites in the read of interest}

\item{cpg}{position of the CpG site of interest}
}
\value{
positions of the CpG sites that are at most WINDOW.SIZE away from one
					another
}
\description{
compute.discordant
This function restricts a read to those CpG sites that are at most WINDOW.SIZE away from one
another.
}
\author{
Michael Scherer
}
