% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RnBeads.R
\name{rnb.calculate.fdrp}
\alias{rnb.calculate.fdrp}
\title{rnb.calculate.fdrp}
\usage{
rnb.calculate.fdrp(rnb.set, bam.path, log.path = getwd(), cores = 1)
}
\arguments{
\item{rnb.set}{RnBSet object containing the CpG sites for which calculation should be conducted}

\item{bam.path}{path to the bam file containing the reads}

\item{log.path}{path to the log directory, if not existing it is created}

\item{cores}{cores available for the analysis}
}
\value{
FDRP scores for the CpG sites in the RnBSet with a higher coverage than COVERAGE.THRESHOLD
}
\description{
This function calculates the FDRP scores for a given RnBSet and a bam file containing the reads.
}
