% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RnBeads.R
\name{rnb.calculate.mhl}
\alias{rnb.calculate.mhl}
\title{rnb.calculate.mhl}
\usage{
rnb.calculate.mhl(rnb.set, bam.path, roi.path = getwd())
}
\arguments{
\item{rnb.set}{RnBSet object containing the CpG sites for which calculation should be conducted}

\item{bam.path}{path to the bam file containing the reads}

\item{roi.path}{path to a directory to which output can be added}
}
\value{
MHL scores for the CpG sites in the RnBSet with a higher coverage than COVERAGE.THRESHOLD
}
\description{
This function calculates the MHL scores for a given RnBSet and a bam file containing the reads.
}
