% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RnBeads.R
\name{rnb.calculate.pdr}
\alias{rnb.calculate.pdr}
\title{rnb.calculate.pdr}
\usage{
rnb.calculate.pdr(rnb.set, bam.path, log.path = getwd(), cores = 1)
}
\arguments{
\item{rnb.set}{RnBSet object containing the CpG sites for which calculation should be conducted}

\item{bam.path}{path to the bam file containing the reads}

\item{log.path}{path to the log directory, if not existing it is created}

\item{cores}{cores available for the analysis}
}
\value{
PDR scores for the CpG sites in the RnBSet with a higher coverage than COVERAGE.THRESHOLD
}
\description{
This function calculates the PDR scores for a given RnBSet and a bam file containing the reads.
}
\examples{
\donttest{
example.rnb.set <- system.file(file.path("extData","small_rnbSet.zip"),package="ISH")
example.bam <- system.file(file.path("extData","small_example.bam"),package="ISH")
fdrp <- rnb.calculate.pdr(rnb.set=example.rnb.set,bam.path=example.bam)
}
}
\author{
Michael Scherer
}
