% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_RnBeads.R
\name{rnb.calculate.qfdrp}
\alias{rnb.calculate.qfdrp}
\title{rnb.calculate.qfdrp}
\usage{
rnb.calculate.qfdrp(
  rnb.set,
  bam.path,
  log.path = getwd(),
  cores = 1,
  use.sex.chromosomes = FALSE
)
}
\arguments{
\item{rnb.set}{RnBSet object containing the CpG sites for which calculation should be conducted}

\item{bam.path}{path to the bam file containing the reads}

\item{log.path}{path to the log directory, if not existing it is created}

\item{cores}{cores available for the analysis}

\item{use.sex.chromosomes}{Flag indicating if scores are also to be computed for the sex chromosomes}
}
\value{
qFDRP scores for the CpG sites in the RnBSet with a higher coverage than COVERAGE.THRESHOLD
}
\description{
This function calculates the qFDRP scores for a given RnBSet and a bam file containing the reads.
}
\examples{
\dontrun{
example.rnb.set <- system.file(file.path("extData","small_rnbSet.zip"),package="WSH")
example.bam <- system.file(file.path("extData","small_example.bam"),package="WSH")
fdrp <- rnb.calculate.qfdrp(rnb.set=example.rnb.set,bam.path=example.bam)
}
}
\author{
Michael Scherer
}
