% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tools.R
\name{run.haplotype.calculation}
\alias{run.haplotype.calculation}
\title{run.haplotype.calculation}
\usage{
run.haplotype.calculation(
  roi,
  bam.file,
  out.folder = getwd(),
  out.name = "hapinfo.txt",
  bam.type = "bismark"
)
}
\arguments{
\item{roi}{Region of interest for which haplotype information should be computed}

\item{bam.file}{path to bam file containing the reads}

\item{out.folder}{folder to write the output}

\item{out.name}{name of the output file}

\item{bam.type}{alignment tools used to create the bam file}
}
\description{
Runs the haplotype calculation associated with MHL's publication
}
\author{
Michael Scherer
}
