% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tools.R
\name{run.methclone}
\alias{run.methclone}
\title{run.methclone}
\usage{
run.methclone(bam.file, out.folder = getwd(), out.name = "methclone")
}
\arguments{
\item{bam.file}{path to bam file containing the reads for defining epialleles}

\item{out.folder}{path to folder where methclones output should be stored}

\item{out.name}{name of the output file}
}
\description{
Runs the methclone software
}
\author{
Michael Scherer
}
