% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tools.R
\name{run.mhl.calculation}
\alias{run.mhl.calculation}
\title{run.mhl.calculation}
\usage{
run.mhl.calculation(hapinfo.file, out.folder = getwd(), out.name = "mhl.txt")
}
\arguments{
\item{hapinfo.file}{path to haplotype information file as produced by \code{\link{run.haplotype.calculation}}}

\item{out.folder}{folder to write the output}

\item{out.name}{name of the output file}
}
\description{
Runs the MHL calculation given the haplotype information file associated with MHL's publication
}
\author{
Michael Scherer
}
