% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{VALID.SCORES}
\alias{VALID.SCORES}
\title{main.R}
\format{
An object of class \code{character} of length 6.
}
\usage{
VALID.SCORES
}
\description{
This scripts contains main functions to calculate the scores.
}
\keyword{datasets}
