% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate.fdrp}
\alias{calculate.fdrp}
\title{calculate.fdrp}
\usage{
calculate.fdrp(
  bam.file,
  anno,
  log.path = getwd(),
  cores = 1,
  window.size = get.option("WINDOW.SIZE"),
  use.sex.chromosomes = FALSE,
  ignore.strand = TRUE
)
}
\arguments{
\item{bam.file}{bath to the bam file to be analyzed
already aligned to a reference genome}

\item{anno}{annotation as a GRanges object with the CpG sites to be analyzed}

\item{log.path}{location of the log file}

\item{cores}{number of cores available for the analysis}

\item{window.size}{window size used to restrict the concordance/discordance classification of each read pair
DEFAULT: 50 as the maximum distance}

\item{use.sex.chromosomes}{Flag indicating if scores are also to be computed for the sex chromosomes}

\item{ignore.strand}{The \code{ignore.strand} parameter from the function \code{\link{findOverlaps}}}
}
\value{
FDRP scores for the given annotation.
}
\description{
This function calculates the FDRP scores for the reads given in the bam files
for the CpGs present in the annotation.
}
\author{
Michael Scherer
}
