% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.score.GRanges}
\alias{compute.score.GRanges}
\title{compute.score.GRanges}
\usage{
compute.score.GRanges(bam.file, range, score = "qfdrp")
}
\arguments{
\item{bam.file}{path to bam file containing the reads}

\item{range}{GRanges object containing the annotation}

\item{score}{The ISH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}
}
\value{
data frame containing the annotation and the computed ISH scores
}
\description{
Main function to compute ISH scores based on a input GRanges object and a bam file containing the reads.
}
\examples{
\donttest{
load(system.file(file.path("extData","example_GRanges.RData"),package="ISH"))
example.bam <- system.file(file.path("extData","small_example.bam"),package="ISH")
fdrp <- compute.score.GRanges(bam.file=example.bam,range=example.GRanges)
}
}
\author{
Michael Scherer
}
