% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.score}
\alias{compute.score}
\title{compute.score}
\usage{
compute.score(bam.file, score = "qfdrp", ...)
}
\arguments{
\item{bam.file}{path to bam file containing the reads}

\item{score}{The ISH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}

\item{...}{additional arugment. Either RnBSet, GRanges or empty (only for Epipolymorphism and Entropy)}
}
\value{
data frame containing the annotation and the computed ISH scores
}
\description{
Generic function to call, passes its arugments either to \code{\link{compute.score.rnb}} or
\code{\link{compute.score.GRanges}}.
}
\author{
Michael Scherer
}
