% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{compute.score.rnb}
\alias{compute.score.rnb}
\title{compute.score.rnb}
\usage{
compute.score.rnb(
  bam.file,
  rnb.set,
  score = "qfdrp",
  use.sex.chromosomes = use.sex.chromosomes
)
}
\arguments{
\item{bam.file}{path to bam file containing the reads}

\item{rnb.set}{path to RnBSet contaning methylation, coverage and sample meta information}

\item{score}{The WSH score which should be computed, needs to be one of \code{fdrp},\code{qfdrp},\code{pdr},\code{epipolymorphism},
\code{entropy} or \code{mhl}}

\item{use.sex.chromosomes}{Flag indicating if scores are also to be computed for the sex chromosomes}
}
\value{
data frame containing the annotation and the computed WSH scores
}
\description{
Main function to compute WSH scores based on a input RnBSet and a bam file containing the reads.
}
\examples{
\dontrun{
example.rnb.set <- system.file(file.path("extData","small_rnbSet.zip"),package="WSH")
example.bam <- system.file(file.path("extData","small_example.bam"),package="WSH")
fdrp <- compute.score.rnb(bam.file=example.bam,rnb.set=example.rnb.set)
}
}
\author{
Michael Scherer
}
