% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set.option}
\alias{set.option}
\title{set.option}
\usage{
set.option(window.size = 50, mapq.filter = 35, max.reads = 40,
  min.overlap = 35, fdrp.type = "FDRP", coverage.threshold = 10,
  methclone.methylation.diff = 0, perl.path = "/usr/bin/perl",
  samtools.path = "/usr/bin")
}
\arguments{
\item{window.size}{Window around the CpG site of interest to consider in FDRP and qFDRP calculation, the higher the value, the more
likely it is to find heterogeneity}

\item{mapq.filter}{mapq filter used to filter out low quality reads}

\item{max.reads}{Maximum number of reads to be considered in FDRP and qFDRP calculation. The scores compute all pairs, therefore this
is a crucial parameter for the running time of the calculation.}

\item{min.overlap}{Miniumum overlap between two reads to consider it as a read pair in FDRP/qFDRP calculation in bp.}

\item{fdrp.type}{FDRP type to be used: either FDRP or qFDRP}

\item{coverage.threshold}{Coverage Threshold emloyed to select the sites in a RnBSet annotation that fullfill havin a coverage
higher than this threshold}

\item{methclone.methylation.diff}{Methylation difference parameter employed by the methclone software. Only sites are considered
that have a methylation difference higher than this value in the methclone package.}

\item{perl.path}{Path to an installation of perl on the machine}

\item{samtools.path}{path to the directory where samtools is located in your machine}
}
\description{
Change global options for IHS score calculation
}
