% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{set.option}
\alias{set.option}
\title{ALL' <- c('window.size','mapq.filter','max.reads','min.overlap','fdrp.type'),
Window around the CpG site of interest to consider in FDRP and qFDRP calculation, the higher the value, the more
likely it is to find heterogeneity
WINDOW.SIZE' <- 50,
mapq filter used to filter out low quality reads
MAPQ.FILTER' <- 35,
Maximum number of reads to be considered in FDRP and qFDRP calculation. The scores compute all pairs, therefore this
is a crucial parameter for the running time of the calculation.
MAX.READS' <- 40,
Miniumum overlap between two reads to consider it as a read pair in FDRP/qFDRP calculation in bp.
MIN.OVERLAP' <- 35,
FDRP type to be used: either FDRP or qFDRP
FDRP.TYPE' <- 'FDRP')
set.option
Change global options for IHS score calculation}
\usage{
set.option(window.size, mapq.filter, max.reads, min.overlap, fdrp.type,
  coverage.threshold, methclone.methylation.diff)
}
\arguments{
\item{window.size:}{Window around the CpG site of interest to consider in FDRP and qFDRP calculation, the higher the value, the more
likely it is to find heterogeneity}

\item{mapq.filer:}{mapq filter used to filter out low quality reads}

\item{max.reads:}{Maximum number of reads to be considered in FDRP and qFDRP calculation. The scores compute all pairs, therefore this
is a crucial parameter for the running time of the calculation.}

\item{min.overlap:}{Miniumum overlap between two reads to consider it as a read pair in FDRP/qFDRP calculation in bp.}

\item{fdrp.type:}{FDRP type to be used: either FDRP or qFDRP}
}
\description{
ALL' <- c('window.size','mapq.filter','max.reads','min.overlap','fdrp.type'),
Window around the CpG site of interest to consider in FDRP and qFDRP calculation, the higher the value, the more
likely it is to find heterogeneity
WINDOW.SIZE' <- 50,
mapq filter used to filter out low quality reads
MAPQ.FILTER' <- 35,
Maximum number of reads to be considered in FDRP and qFDRP calculation. The scores compute all pairs, therefore this
is a crucial parameter for the running time of the calculation.
MAX.READS' <- 40,
Miniumum overlap between two reads to consider it as a read pair in FDRP/qFDRP calculation in bp.
MIN.OVERLAP' <- 35,
FDRP type to be used: either FDRP or qFDRP
FDRP.TYPE' <- 'FDRP')
set.option
Change global options for IHS score calculation
}
